Rudimentary support for loading mods at a later stage than at launch, too.
This is the precursor to a more full-fletched mod loading overhaul so that a per-game/server mod manager + (auto) mod updating is actually possible. The assumption for now is that unloading/disabling mods will require a game relaunch, though loading up additional mods may work without restart depending on the mod (and at which point the mod is loaded, of course).